backend/requirements.txt
django-cors-headers = "==3.7.0"
djangorestframework = "==3.12.2"
djangorestframework-simplejwt = "==4.6.0"